Made the changes from #2191283
