+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "drupal/mailsystem",
-    "description": "Mailsystem",
+    "description": "Mail System",
     "type": "drupal-module",
     "license": "GPL-2.0+",
     "minimum-stability": "dev",
