+++ mailsystem.routing.yml
@@ -2,6 +2,6 @@ mailsystem.settings:
   path: '/admin/config/system/mailsystem'
   defaults:
     _form: 'Drupal\mailsystem\AdminForm'
-    _title: 'Configure the Mailsystem'
+    _title: 'Configure the Mail System'
   requirements:
     _permission: 'administer_mailsystem'
