+++ tests/modules/mailsystem_test/src/Plugin/Mail/Dummy.php
@@ -1,10 +1,10 @@
 <?php
 /**
-* @file
-* Contains \Drupal\mailsystem\Plugin\mailsystem\Dummy.
-*/
+ * @file
+ * Contains \Drupal\mailsystem_test\Plugin\Mail\Dummy.
+ */
 
-namespace Drupal\mailsystem\Plugin\Mail;
+namespace Drupal\mailsystem_test\Plugin\Mail;
 
 use Drupal\Core\Mail\MailInterface;
 
@@ -44,4 +44,5 @@ class Dummy implements MailInterface {
     ), 'Dummy: mail()');
     return TRUE;
   }
+
 }
